requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10.2
django-extensions==1.7.4
djangorestframework==3.5.1
drf-nested-routers==0.11.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
